+++ Assets/AIFrame/Scripts/Tools/MathHelper.cs
@@ -8,14 +8,14 @@ namespace AIFrame
         /// <summary>
         /// Sigmoid激活函数
         /// </summary>
-        public static double SigmoidFunction(double _xValue)
+        public static float SigmoidFunction(float _xValue)
         {
             if (_xValue > 10) return 1.0f;
             else if (_xValue < -10) return 0.0f;
             else return (float)(1.0 / (1.0 + Math.Exp(-_xValue)));
         }
 
-        public static double SoftSignFunction(double xValue)
+        public static float SoftSignFunction(float xValue)
         {
             return xValue / (1 + Math.Abs(xValue));
         }
